Added tinyscheme.
Supporting keystores and a web of trust will require making more code.
It will also require more user interaction code. The best way to do this
is a Scheme intepreter. I've thought hard about alternatives, but they
don't seem as extensible. Being able to make any tree or list into a human-redable,
understandable form is a big advantage, but parsing in C seemed a bit beyond my
abilities or what I wanted to deal with. When crawling lists, use LISP.

For the purposes of signing keys we will use canonical SEXPRs as defined for SPKI.
This is very much a work in progress and may not lead anywhere.